refactor: rename commit_id_fetcher -> task_info_fetcher
